feat: stable hash idle detection, fractional heartbeat minutes, pin=false new-message behavior
- heartbeat_mins/heartbeat_interval changed from u64 to f64 (config.rs, cli.rs,
  tmux_wrapper.rs, main.rs); effective_heartbeat_secs() replaces
  effective_heartbeat_mins() in tmux.rs
- content_hash_stable: use saturating_sub(8) to exclude animated TUI bottom
  (status bar, Bramblesnit mascot) from stable hash — fixes OMC/Claude idle
  detection false positives
- add count_changed_lines() + STABLE_CHANGE_THRESHOLD=4: only update
  last_stable_change when substantial content changes (>4 lines differ), ignoring
  animation frames in both wrapper and config pane update paths
- remove last_heartbeat = Some(now) from content-changed branch in wrapper path;
  heartbeat timing is now managed entirely via last_stable_change in TmuxPaneState
- discord: when pin_status/pin_summary/pin_alerts/pin_keywords is false, emit new
  message each time instead of editing keyed slot
- summarize: fix UTF-8 panic in truncate_for_summarizer on multi-byte chars

+++ src/cli.rs
@@ -461,9 +461,9 @@ pub struct TmuxNewArgs {
     pub retry_enter_delay_ms: u64,
     #[arg(long)]
     pub shell: Option<String>,
-    /// Send a heartbeat notification every N minutes (0 = disabled).
-    #[arg(long, default_value_t = 1)]
-    pub heartbeat_mins: u64,
+    /// Send a heartbeat notification every N minutes (0 = disabled). Supports fractional values (e.g. 0.333 ≈ 20s).
+    #[arg(long, default_value_t = 1.0)]
+    pub heartbeat_mins: f64,
     /// Detect when the session is waiting for user input.
     #[arg(long, default_value_t = true, action = ArgAction::Set)]
     pub detect_waiting: bool,
@@ -514,9 +514,9 @@ pub struct TmuxWatchArgs {
     pub format: Option<TmuxWrapperFormat>,
     #[arg(long, default_value_t = true, action = ArgAction::Set)]
     pub retry_enter: bool,
-    /// Send a heartbeat notification every N minutes (0 = disabled).
-    #[arg(long, default_value_t = 1)]
-    pub heartbeat_mins: u64,
+    /// Send a heartbeat notification every N minutes (0 = disabled). Supports fractional values (e.g. 0.333 ≈ 20s).
+    #[arg(long, default_value_t = 1.0)]
+    pub heartbeat_mins: f64,
     /// Detect when the session is waiting for user input.
     #[arg(long, default_value_t = true, action = ArgAction::Set)]
     pub detect_waiting: bool,
